수정 음식데이터 테스트 fixtures

+++ scripts/remove_server.sh
@@ -37,16 +37,23 @@ if [ -e $PROJECT_PATH ]; then
 fi
 
 # 서비스 중지
-if [ -z "$SERVICE_GUNICONR" ]; then
-    sudo supervisorctl stop gunicorn
-fi
+# if [ -z "$SERVICE_GUNICONR" ]; then
+sudo supervisorctl stop gunicorn
+# fi
 
 if [ -z "$SERVICE_NIGNX" ]; then
     sudo service nginx stop
 fi
 
+sudo apt-get clean
 sudo apt autoremove
+sudo apt autoremove --purge
+sudo apt autoremove --purge snapd
 sudo npm cache clean -f
+sudo mv /var/lib/apt/lists /var/lib/apt/lists.old  
+npm cache clean --force
 
-sudo swapoff -v /swapfile
-sudo rm /swapfile
+if [ -e /swapfile ]; then
+    sudo swapoff -v /swapfile
+    sudo rm /swapfile
+fi
